#781 [REFACTOR] 게시글 리스트 페이지와 게시판 선택 페이지 폴더를 분리
