--- src/pages/BoardPage/BoardListPageHasCategory/BoardListPageHasCategory.module.css
@@ -1,103 +1,0 @@
-body {
-  -ms-overflow-style: none;
-  scrollbar-width: none;
-}
-::-webkit-scrollbar {
-  display: none;
-}
-
-.container {
-  width: 100%;
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  font-size: 0.5rem;
-}
-
-.top {
-  width: 100%;
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  font-size: 0.5rem;
-  padding: 0.625rem 1.25rem;
-  gap: 0.5625rem;
-}
-
-.pencil_icon {
-  display: flex;
-  justify-content: end;
-  width: 100%;
-  max-width: 28.125rem;
-  position: fixed;
-  bottom: 7.5rem;
-  z-index: 1;
-}
-
-.side_menu_btn {
-  width: 1.25rem;
-  height: 0.9375rem;
-  display: flex;
-}
-
-.notification_bar {
-  width: 100%;
-  background-color: var(--pink-1);
-  height: 2.188rem;
-  font-size: 0.688rem;
-  color: var(--pink-3);
-  border-radius: 0.3125rem;
-  display: flex;
-  padding: 0.688rem 0.938rem;
-  align-items: center;
-  gap: 0.861rem;
-}
-
-.notification_icon {
-  width: 0.663rem;
-  height: 0.791rem;
-  display: flex;
-}
-
-.keyword_box {
-  display: flex;
-  width: 100%;
-  justify-content: baseline;
-}
-
-.keyword {
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  padding: 0.1875rem 0.875rem;
-  border-radius: 1.25rem;
-  background: var(--blue-0);
-  margin-right: 0.3125rem;
-  color: var(--blue-3);
-
-  p {
-    font-size: 0.6875rem;
-  }
-}
-
-.content {
-  width: 100%;
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  font-size: 0.5rem;
-  /* 양옆 여백 */
-  padding: 0.25rem 1rem 12rem 1rem;
-  row-gap: 0.5rem;
-}
-
-.sponsor {
-  display: flex;
-  width: 100%;
-  background-size: cover;
-  position: fixed;
-  justify-content: center;
-  bottom: 5.1875rem;
-  /* background-color: var(--white); */
-  padding-bottom: 1.25rem;
-}
