+++ src/pages/BoardPage/index.js
@@ -1,3 +1,1 @@
-export { default as BoardPage } from './BoardPage/BoardPage';
-export { default as BoardListPage } from './BoardListPage/BoardListPage';
-export { default as BoardPostList } from './BoardListPage/BoardPostList/BoardPostList.jsx';
+export { default as BoardPage } from './BoardPage';
