+++ src/pages/PostsPage/index.js
@@ -1,0 +1,3 @@
+export { default as PostsPage } from './PostsPage';
+
+export { default as Posts } from './components/Posts/Posts';
